chat function now working, seperated chat room into new component Room
